+++ src/main/java/ca/tyny/urlshortener/infra/config/SecurityConfig.java
@@ -3,6 +3,7 @@ package ca.tyny.urlshortener.infra.config;
 import ca.tyny.urlshortener.infra.security.CustomUserDetailsService;
 import ca.tyny.urlshortener.infra.security.JwtAuthenticationFilter;
 import ca.tyny.urlshortener.infra.config.properties.SecurityProperties;
+import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -44,6 +45,9 @@ public class SecurityConfig {
                         .sessionCreationPolicy(SessionCreationPolicy.STATELESS))
                 .authenticationProvider(authenticationProvider())
                 .addFilterBefore(jwtAuthFilter, UsernamePasswordAuthenticationFilter.class)
+                .exceptionHandling(ex -> ex
+                        .authenticationEntryPoint((request, response, authException) ->
+                                response.sendError(HttpServletResponse.SC_UNAUTHORIZED)))
                 .authorizeHttpRequests(auth -> auth
                         // Public Endpoints
                         .requestMatchers("/api/v1/auth/**").permitAll() // Login & Register